pyproject.toml
[tool.poetry]
name = "vk-bot"
version = "3.0.1"
description = ""
authors = ["MaHryCT3 <mahryct123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
vkbottle = {git = "https://github.com/vkbottle/vkbottle.git"}
pydantic = "^1.10.7"
aiohttp = "^3.8.4"
loguru = "^0.7.0"
python-dateutil = "^2.8.2"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.261"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
select = ['Q', 'F', 'I', 'ANN', 'B', 'C4', 'TID', 'ARG', 'PLR']
ignore = ['ANN101', 'ANN204', 'ANN401', 'ANN002', 'ANN003', 'PLR0913']  
# ANN - https://beta.ruff.rs/docs/rules/#flake8-annotations-ann, 
# PLR0913 - Too many arguments


[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"] # Unused import
"tests/*" = ['ANN']  # Annotations
"app/handlers/*" = ['ARG001']
"app/core/custom_rules/*" = ['ARG002']

[tool.ruff.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"

[tool.black]
skip-string-normalization = true
exclude = '(\.git|\.venv|\.vscode)'
line-length = 120